requirements.txt
playwright==1.55.0
pytest==8.4.2
pytest-playwright==0.7.1
allure-pytest==2.15.0
python-dotenv==1.2.1
faker==37.12.0